+++ test-cypress/cypress.json
@@ -1,6 +1,7 @@
 {
   "projectId": "dyez6t",
-  "cacophony-api-server": "https://api-test.cacophony.org.nz",
-
-  "integrationFolder": "cypress/api-tests"
+  "env" : {
+    "cacophony-api-server": "https://api-test.cacophony.org.nz", 
+    "testUsers" : {}
+  }
 }
