--- test-cypress/cypress/support/index.js
@@ -1,10 +1,0 @@
-// ***********************************************************
-// This example support/index.js is processed and
-// loaded automatically before each test file.
-//
-// This is a great place to for global configuration.
-// ***********************************************************
-
-import '../commands/api/users';
-import '../commands/api/camera';
-
