+++ test-cypress/cypress/support/index.ts
@@ -1,0 +1,5 @@
+
+// load the global Cypress types
+/// <reference types="cypress" />
+
+import '../commands/api/users';
