+++ test-cypress/package.json
@@ -11,16 +11,14 @@
     "lint:fix": "eslint ./cypress --fix --ext .js,.vue"
   },
   "dependencies": {
-    "cypress": "^3.8.3"
+    "@types/node": "^14.14.25",
+    "cypress": "^6.4.0",
+    "typescript": "^4.1.3"
   },
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-plugin-cypress": "^2.7.0",
     "eslint-plugin-no-only-tests": "^2.4.0"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/TheCacophonyProject/integration-tests.git"
   },
   "author": "",
   "license": "ISC",
