+++ test-cypress/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "target": "es5",
+            "lib": ["es5", "dom"],
+                "types": ["cypress"]
+    },
+    "include": [
+        "**/*.ts", "cypress/api-tests/visits.js"
+    ]
+}
